Added functions DTP_GenericPols_r and DTP_GenericPols_rs
Both functions are called with a natural number n as the only
argument.
The polynomials are returned in lists of GAP polynomials in
indeterminates x1, ..., xn, y1, ..., yn, ti_j_k,
1 <= i < j < k <= n.
Note that parts of the code changed and in the current state
of the code the old functions for computing a DTObj DO NOT
work anymore.